+++ app/Http/Controllers/Api/V1/UserController.php
@@ -30,14 +30,11 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            // policy
-             $this->isAble('store', User::class);
- 
-             return new UserResource(User::create($request->mappedAttributes()));
-         } catch (AuthorizationException $ex) {
-             return $this->error('You are not authorized to create that resource', 401);
-         }
+        if ($this->isAble('store', User::class)) {
+            return new UserResource(User::create($request->mappedAttributes()));
+        }
+
+        return $this->error('You are not authorized to create that resource', 401);
     }
 
     /**
@@ -60,16 +57,16 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('update', $user);
-            
-            $user->update($request->mappedAttributes());
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
     
-            return new UserResource($user);
+                return new UserResource($user);
+            }
+            
+            return $this->error('You are not authorized to update that resource', 401);
+            
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
     }
 
@@ -78,13 +75,14 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('replace', $user);
+            if ($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-    
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
         }
@@ -98,12 +96,14 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('delete', $user);
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
 
-            $user->delete();
+                return $this->ok('User successfully deleted');
+            }
 
-            return $this->ok('User successfully deleted');
+            return $this->error('You are not authorized to delete that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot found.', 404);
         }
